requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
gunicorn==21.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Alembic==1.13.1
snowflake-connector-python==3.7.0
cryptography==41.0.7
requests==2.31.0

